Fixup build after merge
