+++ include/RAJA/policy/openmp/policy.hpp
@@ -117,8 +117,8 @@ struct omp_taskgraph_interval_segit
 ///
 /// Policies for applying OpenMP clauses in forallN loop nests.
 ///
-struct omp_collapse_nowait_exec {
-};
+struct omp_collapse_nowait_exec : 
+  public RAJA::make_policy_pattern<RAJA::Policy::openmp, RAJA::Pattern::forall> {};
 
 ///
 ///////////////////////////////////////////////////////////////////////
@@ -134,8 +134,6 @@ struct omp_reduce : public RAJA::make_policy_pattern<RAJA::Policy::openmp,
 struct omp_reduce_ordered : public omp_reduce {
 };
 
-struct omp_collapse_nowait_exec : 
-  public RAJA::make_policy_pattern<RAJA::Policy::openmp, RAJA::Pattern::forall> {};
 
 }  // closing brace for RAJA namespace
